Module Iota.Main

module Manip : Program.Manip with type t = Stdlib.String.t

module Structure : Structure.Structure with type t = Stdlib.String.t

module Typecheck : Typecheck.Typecheck with type t = Stdlib.String.t

module Guard_inference : Guard_inference.Guard_inference with type t = Stdlib.String.t

type err =

| Menhir

| Type of Typecheck.type_error

| Parse of Program.parse_error

| Structure of Structure.invalidity

| GuardInference of Guard_inference.err

val print_err : err -> unit

val config : Stdlib.Lexing.lexbuf -> Config.config

val main : Config.config -> Stdlib.Lexing.lexbuf -> (string Final_def.final_program, err) Stdlib.result

Provided a configuration and a Lexbuf to parse, this function compiles a Iota code. It:

Parses it using Menhir.

Typechecks it (with Typecheck).

Compiles it to a much simpler final object (with Final).

Tries to infer some information (with Guard_inference)

Ensures its validty (with Structure).