Module Iota__Final

This module compiles a typechecking program to a final one. It

Inlines all let definitions.

Compiles TLink to the corresponding list of Link (taking into account links constraint).

Simplifies Link(x.f) into x.

type err =

| UnknownType of string

| EmptyTLink

val string_of_err : err -> string

val final_of_program : config:Iota.Config.config -> string Iota.Typecheck.typed_program -> (string Iota.Final_def.pre_final_program, err) Stdlib.result

Transform a program that typechecks into a final one