Module Iota__Main

module Manip : Iota.Program.Manip with type t = Stdlib.String.t

module Structure : Iota.Structure.Structure with type t = Stdlib.String.t

module Typecheck : Iota.Typecheck.Typecheck with type t = Stdlib.String.t

module Gfinal : Iota.Gfinal.Guard_inference with type t = Stdlib.String.t

type err =

| Menhir

| Type of Typecheck.type_error

| Parse of Iota.Program.parse_error

| Structure of Structure.invalidity

| GuardInference of Gfinal.err

val print_err : err -> unit

val config : Stdlib.Lexing.lexbuf -> Iota.Config.config

val main : Iota.Config.config -> Stdlib.Lexing.lexbuf -> (string Iota.Final.final_program, err) Stdlib.result

Provided a configuration and a Lexbuf to parse, this function compiles a Iota code. It:

Parses it using Menhir.

Typechecks it (with Typecheck).

Compiles it to a much simpler final object (with Final).

Ensures its validty (with Structure).